requirements.txt
# ML Challenge 2025 - Requirements
# Python 3.12+ compatible

# Core ML/DL libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Gradient Boosting
lightgbm>=4.0.0

# NLP Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0

# Image Processing
Pillow>=10.0.0
requests>=2.31.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (for local development)
jupyter>=1.0.0
ipykernel>=6.25.0
